% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolarGeolocation.R
\name{getTemplateCalib}
\alias{getTemplateCalib}
\title{Calibration function}
\usage{
getTemplateCalib(tagdata, twl, lon, lat, bin = 1, adjust = 300,
  plot = TRUE)
}
\arguments{
\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)
and light levels recorded by the tag at known location.}

\item{twl}{a dataframe with columns \code{Twilight} and
\code{Rise} that are the sequence of defined sunrise and sunset times.}

\item{lon}{calibration longitude.}

\item{lat}{calibration latitude.}

\item{bin}{the grouping factor of light recordings across zenith angles.}

\item{adjust}{timing adjustment for sunset intervals.}

\item{plot}{logical, if \code{TRUE} a plot will be shown.}
}
\value{
a dataframe with one row for each bin of zenith angles. Columns provide minimum and maximum zenith
angle (range), maximum light value (the template) and the rate and scale paramters of a fitted gamma distribution.
}
\description{
Creates a calibration table
}
\details{
A central function for the location estimation process. This function creates a table
with estimated maximum light template as well as gamma distribution parameters for the
recorded light across intervals (bins) of zenith angles.

Note: Plase refer to publication in Wader Study for detailed information (Lisovski 2018).

The \code{adjust.interval} argument can be used to specify a
timing adjustment for tags that report the maximum light level
observed in the preceeding sampling interval. If this argument is
zero, no adjustment is made, otherwise the timestamps of sunset
intervals will be adjusted by this interval to compensate for the
recording behaviour of the tag.
}
