% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolarGeolocation.R
\name{makeMask}
\alias{makeMask}
\title{Spatial mask/grid}
\usage{
makeMask(centre, radius, res = 150, mask = "land", map = NULL,
  plot = FALSE)
}
\arguments{
\item{centre}{centre (lon, lat) of the mask.}

\item{radius}{radius around the centre (in km).}

\item{res}{resolution of grid cells in km, defaults to 150}

\item{mask}{set mask for areas animals are assumed to be restricted to, "sea", "land", or "none"}

\item{map}{the map (spatial polygon) that will be used to define land and sea. If `NULL` the dataset `wrld_simpl` from the package `maptools` will be used.}
}
\value{
raster object giving the locations the animal may have visited
}
\description{
Creates a spatial mask/grid
}
\details{
This function creates a rectangular grid using an equal area projection around a defined centre.
Furthermore it uses defines boundaries and probabilities for possible locations (e.g. land vs. ocean).
}
